add functions listing-align[-element]
Previously the element lines and the header-line were aligned in the
functions `listing-format-[element|header]' respectively.  Doing it
in a seperate function is slightly slower but makes these functions
a bit cleaner and allow us to use function `listing-map-lines' which
would be rather pointless otherwise.

Of course this ain't the real reason for doing this - in the futur
we want to hide/unhide columns on the fly which makes it necessary
to seperate creating and aligning the element lines.