remove function listing-catgorize
This was designed to to do the intial invisibility setup as well as
categorize elements on-demand in cases where we don't want to do it
during the initial setup because of efficiency reasons.

However the last to commits make this function unnecessary.  Also
this function is broken in it's current (and final) form.