prepare for deprecation of this library
`tabulated-list' which will be part of 24.1 appears to be a suitable
replacement for this library.  It does lack some features but instead
of maintaining this alternative library I will build on top of this
standard library.

There will be a final release of `listing' to allow it to die in dignity.