drop support for hiding columns and rows (elements)
Normal movement was extremely slow when many elements were hidden so this
feature was not used.  Hiding columns didn't have these performance issues
but without the possibility to hide rows also it doesn't make much sense
to implement this feature using invisible text.  Removing this also makes
it possible to drop the isearch hacks.  But yes the ability to hide
columns is gone for now.